+++ packages/epics/src/agreements/components/index.ts
@@ -3,4 +3,4 @@ export * from './agreement-card';
 export * from './agreements-section';
 export * from './agreement-detail';
 export * from './agreement-head';
-export * from './create-agreement-form';
+export * from './create-agreement-base-fields';
